+++ CHANGELOG.md
@@ -69,6 +69,7 @@
     and will use the last instance in the list if duplicates are present. Whilst duplicates are technically valid in the lexicon,
     they are not expected to be present in the wild, and this change prevents deserialization errors when they are encountered.
   * `BlueskyAgent.PutPreference` has been removed, use `BlueskyAgent.PutPreferences` instead, which accepts a `Preferences` instance and replaces the entire preferences for the authenticated user.
+* `ActivitySubscription` has been replace with `Actor.ProfileAssociatedActivitySubscription` to match the lexicon definition. `ProfileAssociated` has been updated to use the new record.
 
 ### Fixed